refactor: extract games_turns.py from games.py
Move list_turns and submit_turn handlers to games_turns.py (317 lines).

games.py: 865 → 596 lines (-269, -31.1%)
Cumulative: 1610 → 596 (games.py) + 79 + 374 + 371 + 317 (extracted)
